refactor: movint i18n hook to context
