+++ src/app/(frontend)/(auth)/confirm-signup/page.tsx
@@ -4,7 +4,7 @@ import { useEffect, useReducer } from "react";
 
 import BackLink from "@/components/v1/BackLink";
 import Spinner from "@/components/v1/Spinner";
-import { useI18n } from '@/hooks/useI18n';
+import { useI18n } from "@/contexts/i18nContext";
 import { supabase } from "@/libs/supabase/client";
 import AuthService from "@/services/auth";
