+++ src/app/(frontend)/(auth)/forgot-password/page.tsx
@@ -5,7 +5,7 @@ import { useReducer } from "react";
 import BackLinkComponent from "@/components/v1/BackLink";
 import ButtonComponent from "@/components/v1/Button";
 import InputComponent from "@/components/v1/Input";
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 import { supabase } from "@/libs/supabase/client";
 import AuthService from "@/services/auth";
 import { isValidEmail } from "@/utils/isValidEmail";
