+++ src/app/(frontend)/(auth)/new-password/page.tsx
@@ -8,7 +8,7 @@ import BackLinkComponent from "@/components/v1/BackLink";
 import ButtonComponent from "@/components/v1/Button";
 import InputComponent from "@/components/v1/Input";
 import PasswordStrengthIndicator from "@/components/v1/PasswordStrength";
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 import { supabase } from "@/libs/supabase/client";
 import AuthService from "@/services/auth";
