+++ src/app/(frontend)/(auth)/signin/page.tsx
@@ -11,7 +11,7 @@ import FooterAuthScreenComponent from "@/components/v1/FooterAuthScreen";
 import InputComponent from "@/components/v1/Input";
 import OAuth from "@/components/v1/OAuth";
 import { ROUTES } from "@/constants/routes-constants";
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 import { isValidEmail } from "@/utils/isValidEmail";
 
 const initialState = {
