+++ src/app/(frontend)/(home)/_sections/FaqSection.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useI18n } from '@/hooks/useI18n';
+import { useI18n } from "@/contexts/i18nContext";
 
 const FaqSection = () => {
   const { translate } = useI18n();
