+++ src/app/(frontend)/(home)/_sections/FeaturesSection.tsx
@@ -2,7 +2,8 @@
 
 import { CheckCircleIcon } from "@heroicons/react/24/outline";
 
-import { useI18n } from '@/hooks/useI18n';
+import { useI18n } from "@/contexts/i18nContext";
+
 
 export default function FeaturesSection() {
   const { translate } = useI18n();
