+++ src/app/(frontend)/(home)/_sections/Footer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { ROUTES } from "@/constants/routes-constants";
-import { useI18n } from "@/hooks/useI18n"
+import { useI18n } from "@/contexts/i18nContext";
 
 type FooterProps = {
   isDashboard?: boolean;
