+++ src/app/(frontend)/(home)/_sections/TestimonialSection.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { useI18n } from '@/hooks/useI18n';
+import { useI18n } from "@/contexts/i18nContext";
 
 const TestimonialSection = () => {
   const { translate } = useI18n();
