+++ src/components/v1/Dashboard/Menu.tsx
@@ -3,7 +3,7 @@
 import { LockClosedIcon } from "@heroicons/react/24/solid";
 import { useEffect, useState } from "react";
 
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 import { useModal } from "@/hooks/useModal";
 
 type Tab = {
