+++ src/components/v1/Dashboard/Modal.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import { useI18n } from "@/hooks/useI18n";
+
+import { useI18n } from "@/contexts/i18nContext";
 
 import { Modal as ModalComponent } from "../Modal";
