+++ src/components/v1/Dashboard/Navbar/Notification/index.tsx
@@ -3,7 +3,7 @@
 import { BellIcon } from "@heroicons/react/24/outline";
 import { useState, useEffect, useRef, useCallback } from "react";
 
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 import { supabase } from "@/libs/supabase/client";
 import AuthService from "@/services/auth";
 import NotificationService, {
