+++ src/components/v1/LanguageSelector.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 
 const LanguageSelector = () => {
   const { locale } = useI18n();
