+++ src/components/v1/ManageBilling.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/navigation";
 
 import { useState } from "react";
 
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 
 import ButtonComponent from "./Button";
