+++ src/components/v1/Navbar.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 
 import { useEffect, useState } from "react";
 
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/contexts/i18nContext";
 import { supabase } from "@/libs/supabase/client";
 import AuthService from "@/services/auth";
