+++ src/components/v1/Pricing/PlanCard.tsx
@@ -1,4 +1,4 @@
-import { useI18n } from '@/hooks/useI18n';
+import { useI18n } from '@/contexts/i18nContext';
 
 import ButtonComponent from '../Button';
