+++ src/components/v1/Pricing/index.tsx
@@ -3,9 +3,9 @@
 import { useState } from 'react';
 
 import { HAS_FREE_TRIAL } from '@/constants/has-free-trial';
+import { useI18n } from '@/contexts/i18nContext';
 import { useCheckout } from '@/hooks/useCheckout';
 import { useFetchPlans } from '@/hooks/useFetchPlans';
-import { useI18n } from '@/hooks/useI18n';
 
 import Spinner from '../Spinner';
 import PlanCard from './PlanCard';
