+++ src/contexts/i18nContext.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { createContext, ReactNode } from "react";
+import { createContext, ReactNode, useContext } from "react";
 
 export interface I18nContextProps {
   locale: string;
@@ -43,3 +43,25 @@ export const I18nProvider = ({
     </I18nContext.Provider>
   );
 };
+
+export const useI18n = (basePath?: string) => {
+  const context = useContext(I18nContext);
+  if (!context) {
+    throw new Error("useI18n must be used within an I18nProvider");
+  }
+
+  const { translate } = context;
+
+  type TFunction = typeof translate;
+
+  const scopedT: TFunction = (key: string) => {
+    const fullKey = basePath ? `${basePath}.${key}` : key;
+    return translate(fullKey);
+  };
+
+  return {
+    ...context,
+    translate: scopedT,
+  };
+};
+
