+++ src/hooks/useFetchPlans.ts
@@ -3,8 +3,8 @@ import { useState, useEffect } from "react";
 import { Plan } from "@/components/v1/Pricing/PlanCard";
 import { FIXED_CURRENCY } from "@/constants/fixed-currency";
 import { HAS_FREE_TRIAL } from "@/constants/has-free-trial";
+import { useI18n } from "@/contexts/i18nContext";
 
-import { useI18n } from "./useI18n";
 
 export const useFetchPlans = (
   hasFreeplan: boolean,
